refactor(forge): rehome GitHub services into plugins/builtin/github
- Move 22 GitHub service files from electron/services/github/ to
  plugins/builtin/github/main/, including __tests__/
- Add plugin.json declaring daintree.github with github forge provider
  contribution (matches: github.com, capabilities: issues, pulls,
  reviews, required-checks, draft-prs, assignees, releases)
- Add forgeProvider.ts implementing ForgeProviderImpl (parseRemote,
  list/get for issues+PRs, findPRByBranch, getCIStatus,
  getRepoMetadata, getRateLimit, reviews.getReviewThreads, URL
  builders, getCredentials, validateCredentials)
- Add activate(host) that calls host.registerForgeProvider
- Add compat shim files at electron/services/github/{index,GitHubAuth,
  GitHubQueries,GitHubTokenHealthService}.ts that re-export from the
  plugin — keeps window.electron.github.* IPC channels and existing
  vi.mock() test paths working until #8063 lands
- Wire build:main esbuild to bundle plugin source in same run with
  splitting:true so singletons (rate-limit service, token health,
  auth) share a single chunk across electron main and plugin bundles
- Update outdir/outbase so plugin output lands at
  dist-electron/plugins/builtin/github/main/index.js while electron
  main output paths stay unchanged
- Copy plugin.json next to the compiled output via
  copyBuiltInPluginManifests()
- Repoint PluginService.getBuiltinDir() dev path to dist-electron/
  plugins/builtin and electron-builder extraResources to match
- Expand electron tsconfig rootDir/include to cover the plugin source
  so the composite project keeps typechecking the rehomed files
- Allow vitest to find tests under plugins/

Refs #8060
