+++ plugins/builtin/github/main/forgeProvider.ts
@@ -1,0 +1,522 @@
+import type { GraphQlQueryResponseData } from "@octokit/graphql";
+import type {
+  AuthValidation,
+  CIStatus,
+  Credentials,
+  ForgeProviderImpl,
+  ForgeUser,
+  ForgeLabel,
+  Issue,
+  ListOptions,
+  NormalizedIssueState,
+  NormalizedPRState,
+  PR,
+  Page,
+  RateLimitInfo,
+  RepoMetadata,
+  RepoRef,
+  ReviewCapability,
+  ReviewThread,
+} from "../../../../shared/types/forge.js";
+import { GitHubAuth, GITHUB_API_TIMEOUT_MS } from "./GitHubAuth.js";
+import { validateGitHubToken } from "./GitHubToken.js";
+import { parseGitHubRepoUrl } from "./GitHubRepoContext.js";
+import {
+  LIST_ISSUES_QUERY,
+  LIST_PRS_QUERY,
+  SEARCH_QUERY,
+  GET_ISSUE_QUERY,
+  GET_PR_QUERY,
+  GET_PR_REVIEW_THREADS_QUERY,
+} from "./GitHubQueries.js";
+import { gitHubRateLimitService } from "./GitHubRateLimitService.js";
+import { parseGitHubError } from "./GitHubErrors.js";
+import { deriveRequiredCIStatus, normalizeRawState } from "./prRequiredCIStatus.js";
+import type { RollupContextNode } from "./prRequiredCIStatus.js";
+
+const REPO_METADATA_QUERY = `
+  query GetRepoMetadata($owner: String!, $repo: String!) {
+    repository(owner: $owner, name: $repo) {
+      defaultBranchRef { name }
+      isPrivate
+      isFork
+      isArchived
+      description
+      licenseInfo { name }
+      repositoryTopics(first: 20) { nodes { topic { name } } }
+    }
+    rateLimit { cost remaining resetAt }
+  }
+`;
+
+const PR_CI_STATUS_QUERY = `
+  query GetPRCIStatus($owner: String!, $repo: String!, $number: Int!) {
+    repository(owner: $owner, name: $repo) {
+      pullRequest(number: $number) {
+        commits(last: 1) {
+          nodes {
+            commit {
+              statusCheckRollup {
+                state
+                contexts(first: 100) {
+                  nodes {
+                    __typename
+                    ... on CheckRun {
+                      name
+                      conclusion
+                      status
+                      isRequired(pullRequestNumber: $number)
+                    }
+                    ... on StatusContext {
+                      context
+                      state
+                      isRequired(pullRequestNumber: $number)
+                    }
+                  }
+                }
+              }
+            }
+          }
+        }
+      }
+    }
+    rateLimit { cost remaining resetAt }
+  }
+`;
+
+function requireClient(): NonNullable<ReturnType<typeof GitHubAuth.createClient>> {
+  const client = GitHubAuth.createClient();
+  if (!client) {
+    throw new Error("GitHub token not configured. Set it in Settings.");
+  }
+  return client;
+}
+
+function isoToMs(value: unknown): number {
+  if (typeof value !== "string") return 0;
+  const t = Date.parse(value);
+  return Number.isFinite(t) ? t : 0;
+}
+
+function isoToMsOrNull(value: unknown): number | null {
+  if (typeof value !== "string") return null;
+  const t = Date.parse(value);
+  return Number.isFinite(t) ? t : null;
+}
+
+function toForgeUser(node: unknown): ForgeUser | undefined {
+  if (!node || typeof node !== "object") return undefined;
+  const n = node as { login?: unknown; avatarUrl?: unknown };
+  if (typeof n.login !== "string") return undefined;
+  return {
+    login: n.login,
+    ...(typeof n.avatarUrl === "string" ? { avatarUrl: n.avatarUrl } : {}),
+    rawData: node,
+  };
+}
+
+function toForgeUsers(node: unknown): ForgeUser[] {
+  const nodes = (node as { nodes?: unknown[] } | undefined)?.nodes;
+  if (!Array.isArray(nodes)) return [];
+  return nodes.map(toForgeUser).filter((u): u is ForgeUser => u !== undefined);
+}
+
+function toForgeLabels(node: unknown): ForgeLabel[] {
+  const nodes = (node as { nodes?: unknown[] } | undefined)?.nodes;
+  if (!Array.isArray(nodes)) return [];
+  const out: ForgeLabel[] = [];
+  for (const n of nodes) {
+    if (!n || typeof n !== "object") continue;
+    const label = n as { name?: unknown; color?: unknown };
+    if (typeof label.name !== "string") continue;
+    out.push({
+      name: label.name,
+      ...(typeof label.color === "string" ? { color: label.color } : {}),
+    });
+  }
+  return out;
+}
+
+function normalizeIssueState(rawState: string): NormalizedIssueState {
+  return rawState.toUpperCase() === "CLOSED" ? "closed" : "open";
+}
+
+function normalizePRState(rawState: string, merged: boolean): NormalizedPRState {
+  if (merged) return "merged";
+  const upper = rawState.toUpperCase();
+  if (upper === "CLOSED") return "closed";
+  if (upper === "MERGED") return "merged";
+  return "open";
+}
+
+function toForgeIssue(node: Record<string, unknown>): Issue {
+  const rawState = typeof node.state === "string" ? node.state : "OPEN";
+  return {
+    number: node.number as number,
+    title: (node.title as string) ?? "",
+    body: (node.bodyText as string) ?? "",
+    state: normalizeIssueState(rawState),
+    rawState,
+    url: (node.url as string) ?? "",
+    author: toForgeUser(node.author),
+    assignees: toForgeUsers(node.assignees),
+    labels: toForgeLabels(node.labels),
+    createdAt: isoToMs(node.createdAt ?? node.updatedAt),
+    updatedAt: isoToMs(node.updatedAt),
+    closedAt: isoToMsOrNull(node.closedAt),
+    rawData: node,
+  };
+}
+
+function toForgePR(node: Record<string, unknown>): PR {
+  const merged = node.merged === true;
+  const rawState = typeof node.state === "string" ? node.state : "OPEN";
+  return {
+    number: node.number as number,
+    title: (node.title as string) ?? "",
+    body: (node.bodyText as string) ?? "",
+    state: normalizePRState(rawState, merged),
+    rawState,
+    isDraft: node.isDraft === true,
+    merged,
+    url: (node.url as string) ?? "",
+    author: toForgeUser(node.author),
+    baseRef: (node.baseRefName as string) ?? "",
+    headRef: (node.headRefName as string) ?? "",
+    mergeable: undefined,
+    createdAt: isoToMs(node.createdAt ?? node.updatedAt),
+    updatedAt: isoToMs(node.updatedAt),
+    closedAt: isoToMsOrNull(node.closedAt),
+    mergedAt: isoToMsOrNull(node.mergedAt),
+    rawData: node,
+  };
+}
+
+function mapIssueGraphQLStates(state: ListOptions["state"]): string[] {
+  if (state === "closed") return ["CLOSED"];
+  if (state === "all") return ["OPEN", "CLOSED"];
+  return ["OPEN"];
+}
+
+function mapPRGraphQLStates(state: ListOptions["state"]): string[] {
+  if (state === "closed") return ["CLOSED", "MERGED"];
+  if (state === "all") return ["OPEN", "CLOSED", "MERGED"];
+  return ["OPEN"];
+}
+
+function buildOrderBy(opts: ListOptions): { field: string; direction: string } {
+  const direction = opts.direction === "asc" ? "ASC" : "DESC";
+  const field = opts.sort === "updated" ? "UPDATED_AT" : "CREATED_AT";
+  return { field, direction };
+}
+
+async function runQuery(
+  query: string,
+  variables: Record<string, unknown>
+): Promise<GraphQlQueryResponseData> {
+  const client = requireClient();
+  try {
+    const response = (await client(query, {
+      ...variables,
+      request: { signal: AbortSignal.timeout(GITHUB_API_TIMEOUT_MS) },
+    })) as GraphQlQueryResponseData;
+    gitHubRateLimitService.updateFromGraphQL(response);
+    return response;
+  } catch (error) {
+    throw new Error(parseGitHubError(error));
+  }
+}
+
+async function listIssuesImpl(repo: RepoRef, opts: ListOptions): Promise<Page<Issue>> {
+  const limit = opts.perPage ?? 20;
+  const orderBy = buildOrderBy(opts);
+
+  const response = await runQuery(LIST_ISSUES_QUERY, {
+    owner: repo.owner,
+    repo: repo.repo,
+    states: mapIssueGraphQLStates(opts.state),
+    cursor: opts.cursor ?? null,
+    limit,
+    orderBy,
+  });
+
+  const issues = (response?.repository as Record<string, unknown> | undefined)?.issues as
+    | {
+        nodes?: unknown[];
+        pageInfo?: { hasNextPage?: boolean; endCursor?: string | null };
+        totalCount?: number;
+      }
+    | undefined;
+  const nodes = (issues?.nodes ?? []) as Array<Record<string, unknown>>;
+  return {
+    items: nodes.filter(Boolean).map(toForgeIssue),
+    nextCursor: issues?.pageInfo?.endCursor ?? null,
+    hasMore: issues?.pageInfo?.hasNextPage ?? false,
+    ...(typeof issues?.totalCount === "number" ? { totalCount: issues.totalCount } : {}),
+  };
+}
+
+async function listPRsImpl(repo: RepoRef, opts: ListOptions): Promise<Page<PR>> {
+  const limit = opts.perPage ?? 20;
+  const orderBy = buildOrderBy(opts);
+
+  const response = await runQuery(LIST_PRS_QUERY, {
+    owner: repo.owner,
+    repo: repo.repo,
+    states: mapPRGraphQLStates(opts.state),
+    cursor: opts.cursor ?? null,
+    limit,
+    orderBy,
+  });
+
+  const prs = (response?.repository as Record<string, unknown> | undefined)?.pullRequests as
+    | {
+        nodes?: unknown[];
+        pageInfo?: { hasNextPage?: boolean; endCursor?: string | null };
+        totalCount?: number;
+      }
+    | undefined;
+  const nodes = (prs?.nodes ?? []) as Array<Record<string, unknown>>;
+  return {
+    items: nodes.filter(Boolean).map(toForgePR),
+    nextCursor: prs?.pageInfo?.endCursor ?? null,
+    hasMore: prs?.pageInfo?.hasNextPage ?? false,
+    ...(typeof prs?.totalCount === "number" ? { totalCount: prs.totalCount } : {}),
+  };
+}
+
+async function getIssueImpl(repo: RepoRef, number: number): Promise<Issue | null> {
+  const response = await runQuery(GET_ISSUE_QUERY, {
+    owner: repo.owner,
+    repo: repo.repo,
+    number,
+  });
+  const issue = (response?.repository as Record<string, unknown> | undefined)?.issue as
+    | Record<string, unknown>
+    | null
+    | undefined;
+  return issue ? toForgeIssue(issue) : null;
+}
+
+async function getPRImpl(repo: RepoRef, number: number): Promise<PR | null> {
+  const response = await runQuery(GET_PR_QUERY, {
+    owner: repo.owner,
+    repo: repo.repo,
+    number,
+  });
+  const pr = (response?.repository as Record<string, unknown> | undefined)?.pullRequest as
+    | Record<string, unknown>
+    | null
+    | undefined;
+  return pr ? toForgePR(pr) : null;
+}
+
+async function findPRByBranchImpl(repo: RepoRef, branchName: string): Promise<PR | null> {
+  const searchQuery = `repo:${repo.owner}/${repo.repo} is:pr head:${branchName} sort:created-desc`;
+  const response = await runQuery(SEARCH_QUERY, {
+    searchQuery,
+    type: "ISSUE",
+    cursor: null,
+    limit: 1,
+  });
+  const nodes = ((response?.search as { nodes?: unknown[] } | undefined)?.nodes ?? []) as Array<
+    Record<string, unknown>
+  >;
+  const first = nodes.find(Boolean);
+  return first ? toForgePR(first) : null;
+}
+
+async function getCIStatusImpl(repo: RepoRef, prNumber: number): Promise<CIStatus | null> {
+  const response = await runQuery(PR_CI_STATUS_QUERY, {
+    owner: repo.owner,
+    repo: repo.repo,
+    number: prNumber,
+  });
+
+  const pr = (response?.repository as Record<string, unknown> | undefined)?.pullRequest as
+    | Record<string, unknown>
+    | null
+    | undefined;
+  if (!pr) return null;
+
+  const commits = pr.commits as
+    | { nodes?: Array<{ commit?: { statusCheckRollup?: unknown } }> }
+    | undefined;
+  const rollup = commits?.nodes?.[0]?.commit?.statusCheckRollup as
+    | {
+        state?: string;
+        contexts?: { nodes?: RollupContextNode[]; pageInfo?: { hasNextPage?: boolean } };
+      }
+    | undefined;
+
+  const contextNodes = rollup?.contexts?.nodes ?? null;
+  const hasNextPage = rollup?.contexts?.pageInfo?.hasNextPage === true;
+  const derived = deriveRequiredCIStatus(contextNodes, hasNextPage, rollup?.state ?? null);
+
+  let state: CIStatus["state"] = "unknown";
+  const effective = derived.ciStatus;
+  if (effective === "SUCCESS") state = "success";
+  else if (effective === "FAILURE" || effective === "ERROR") state = "failure";
+  else if (effective === "PENDING" || effective === "EXPECTED") state = "pending";
+  else if (effective === undefined && (derived.ciSummary?.requiredTotal ?? 0) === 0)
+    state = "neutral";
+
+  const total = derived.ciSummary?.requiredTotal ?? 0;
+  const failed = derived.ciSummary?.requiredFailing ?? 0;
+  const pending = derived.ciSummary?.requiredPending ?? 0;
+  const passed = Math.max(0, total - failed - pending);
+  const requiredChecksPassing =
+    derived.ciSummary !== undefined
+      ? derived.ciSummary.requiredTotal > 0 &&
+        derived.ciSummary.requiredFailing === 0 &&
+        derived.ciSummary.requiredPending === 0
+      : undefined;
+
+  return {
+    state,
+    total,
+    passed,
+    failed,
+    pending,
+    ...(requiredChecksPassing !== undefined ? { requiredChecksPassing } : {}),
+    rawData: rollup ?? null,
+  };
+}
+
+async function getRepoMetadataImpl(repo: RepoRef): Promise<RepoMetadata> {
+  const response = await runQuery(REPO_METADATA_QUERY, {
+    owner: repo.owner,
+    repo: repo.repo,
+  });
+
+  const repository = (response?.repository as Record<string, unknown> | undefined) ?? {};
+  const defaultBranch =
+    ((repository.defaultBranchRef as { name?: unknown } | null | undefined)?.name as
+      | string
+      | undefined) ?? "main";
+  const license =
+    ((repository.licenseInfo as { name?: unknown } | null | undefined)?.name as string | null) ??
+    null;
+  const topicNodes = ((
+    repository.repositoryTopics as { nodes?: Array<{ topic?: { name?: unknown } }> } | undefined
+  )?.nodes ?? []) as Array<{ topic?: { name?: unknown } }>;
+  const topics = topicNodes
+    .map((n) => (typeof n.topic?.name === "string" ? n.topic.name : null))
+    .filter((s): s is string => s !== null);
+
+  return {
+    defaultBranch,
+    isPrivate: repository.isPrivate === true,
+    isFork: repository.isFork === true,
+    isArchived: repository.isArchived === true,
+    description: (repository.description as string | null | undefined) ?? null,
+    license,
+    topics,
+    rawData: repository,
+  };
+}
+
+async function getReviewThreadsImpl(repo: RepoRef, prNumber: number): Promise<ReviewThread[]> {
+  const threads: ReviewThread[] = [];
+  let cursor: string | null = null;
+  while (true) {
+    const response = await runQuery(GET_PR_REVIEW_THREADS_QUERY, {
+      owner: repo.owner,
+      repo: repo.repo,
+      number: prNumber,
+      cursor,
+    });
+    const reviewThreads = (
+      (response?.repository as Record<string, unknown> | undefined)?.pullRequest as
+        | {
+            reviewThreads?: {
+              nodes?: unknown[];
+              pageInfo?: { hasNextPage?: boolean; endCursor?: string | null };
+            };
+          }
+        | undefined
+    )?.reviewThreads;
+    const nodes = (reviewThreads?.nodes ?? []) as Array<Record<string, unknown>>;
+    for (const n of nodes) {
+      if (!n) continue;
+      const id = `${repo.owner}/${repo.repo}#${prNumber}:${threads.length}`;
+      threads.push({ id, rawData: n });
+    }
+    if (reviewThreads?.pageInfo?.hasNextPage && reviewThreads.pageInfo.endCursor) {
+      cursor = reviewThreads.pageInfo.endCursor;
+      continue;
+    }
+    break;
+  }
+  return threads;
+}
+
+function getRateLimitImpl(): Promise<RateLimitInfo> {
+  const state = gitHubRateLimitService.getState();
+  if (!state.blocked) {
+    return Promise.resolve({ limit: null, remaining: null, resetAt: null });
+  }
+  return Promise.resolve({
+    limit: null,
+    remaining: 0,
+    resetAt: state.resetAt ?? null,
+    ...(state.kind === "secondary" ? { secondaryThrottled: true } : {}),
+  });
+}
+
+const reviewCapability: ReviewCapability = {
+  getReviewThreads: getReviewThreadsImpl,
+};
+
+export const githubForgeProvider: ForgeProviderImpl = {
+  async getCredentials(): Promise<Credentials | null> {
+    const token = GitHubAuth.getToken();
+    if (!token) return null;
+    return { kind: "bearer", value: token };
+  },
+
+  async validateCredentials(): Promise<AuthValidation> {
+    const token = GitHubAuth.getToken();
+    if (!token) {
+      return { valid: false, error: "No GitHub token configured" };
+    }
+    const result = await validateGitHubToken(token);
+    return {
+      valid: result.valid,
+      scopes: result.scopes,
+      expiresAt: null,
+      ...(result.error ? { error: result.error } : {}),
+    };
+  },
+
+  parseRemote(url: string): RepoRef | null {
+    const parsed = parseGitHubRepoUrl(url);
+    if (!parsed) return null;
+    return {
+      host: "github.com",
+      owner: parsed.owner,
+      repo: parsed.repo,
+      rawData: { url },
+    };
+  },
+
+  listIssues: listIssuesImpl,
+  listPRs: listPRsImpl,
+  getIssue: getIssueImpl,
+  getPR: getPRImpl,
+  findPRByBranch: findPRByBranchImpl,
+  getCIStatus: getCIStatusImpl,
+  getRepoMetadata: getRepoMetadataImpl,
+
+  buildIssueUrl(repo: RepoRef, number: number): string {
+    return `https://github.com/${repo.owner}/${repo.repo}/issues/${number}`;
+  },
+
+  buildPRUrl(repo: RepoRef, number: number): string {
+    return `https://github.com/${repo.owner}/${repo.repo}/pull/${number}`;
+  },
+
+  getRateLimit: getRateLimitImpl,
+
+  reviews: reviewCapability,
+};
