+++ plugins/builtin/github/main/index.ts
@@ -1,0 +1,126 @@
+import type { PluginHostApi } from "../../../../shared/types/plugin.js";
+import { githubForgeProvider } from "./forgeProvider.js";
+
+/**
+ * Plugin activation entry point — called by `PluginService` after manifest
+ * validation. Registers the `github` forge provider declared in `plugin.json`.
+ * The descriptor id MUST match `contributes.forgeProviders[].id` or
+ * `host.registerForgeProvider` throws.
+ */
+export function activate(host: PluginHostApi): () => void {
+  return host.registerForgeProvider({ id: "github" }, githubForgeProvider);
+}
+
+export { githubForgeProvider } from "./forgeProvider.js";
+
+export {
+  GitHubAuth,
+  GITHUB_API_TIMEOUT_MS,
+  GITHUB_AUTH_TIMEOUT_MS,
+  captureAuthMetadata,
+} from "./GitHubAuth.js";
+export type { GitHubTokenConfig, GitHubTokenValidation } from "./GitHubAuth.js";
+
+export {
+  gitHubRateLimitService,
+  GitHubRateLimitError,
+  PRIMARY_RESET_BUFFER_MS,
+} from "./GitHubRateLimitService.js";
+export type { ShouldBlockResult } from "./GitHubRateLimitService.js";
+
+export {
+  gitHubTokenHealthService,
+  GitHubTokenHealthServiceImpl,
+  HEALTH_CHECK_INTERVAL_MS,
+  HEALTH_CHECK_FOCUS_COOLDOWN_MS,
+  HEALTH_CHECK_FETCH_TIMEOUT_MS,
+} from "./GitHubTokenHealthService.js";
+
+export {
+  REPO_STATS_QUERY,
+  REPO_STATS_AND_PAGE_QUERY,
+  PROJECT_HEALTH_QUERY,
+  LIST_ISSUES_QUERY,
+  LIST_PRS_QUERY,
+  SEARCH_QUERY,
+  GET_ISSUE_QUERY,
+  GET_PR_QUERY,
+  GET_PR_REVIEW_THREADS_QUERY,
+  buildBatchPRQuery,
+  buildBatchRequiredChecksQuery,
+  buildGitHubSearchQuery,
+} from "./GitHubQueries.js";
+
+export { deriveRequiredCIStatus } from "./prRequiredCIStatus.js";
+export type { RollupContextNode, DerivedCIResult } from "./prRequiredCIStatus.js";
+
+export type {
+  RepoContext,
+  RepoStats,
+  RepoStatsResult,
+  LinkedPR,
+  PRCheckResult,
+  PRCheckCandidate,
+  BatchPRCheckResult,
+  CIStatus,
+  ProjectHealth,
+  ProjectHealthResult,
+} from "./types.js";
+
+// Token helpers
+export {
+  getGitHubToken,
+  hasGitHubToken,
+  setGitHubToken,
+  clearGitHubToken,
+  getGitHubConfig,
+  getGitHubConfigAsync,
+  validateGitHubToken,
+} from "./GitHubToken.js";
+
+// Token orchestration
+export { setTokenAndSync, clearTokenAndSync } from "./GitHubTokenOrchestrator.js";
+
+// Repo context
+export {
+  parseGitHubRepoUrl,
+  getRepoContext,
+  getRepoInfo,
+  getRepoUrl,
+  getIssueUrl,
+  withRepoContextRetry,
+} from "./GitHubRepoContext.js";
+
+// Cache management
+export { clearGitHubCaches, clearPRCaches } from "./GitHubCaches.js";
+
+// Stats
+export {
+  getRepoStats,
+  getRepoStatsAndPage,
+  getFirstPageCache,
+  getRepoStatsComplete,
+} from "./GitHubStats.js";
+export type { RepoStatsAndPageResult, RepoStatsCompleteResult } from "./GitHubStats.js";
+
+// Project health
+export { getProjectHealth, buildEmptyProjectHealthData } from "./GitHubHealth.js";
+
+// PR discovery
+export { batchCheckLinkedPRs } from "./GitHubPRDiscovery.js";
+
+// Error handling
+export { parseGitHubError } from "./GitHubErrors.js";
+
+// PRs
+export { listPullRequests, getPRByNumber, getPRTooltip, getPRReviewThreads } from "./GitHubPRs.js";
+
+// Issues
+export { listIssues, getIssueByNumber, getIssueTooltip, assignIssue } from "./GitHubIssues.js";
+export type { AssignIssueResult } from "./GitHubIssues.js";
+
+// Rate limit API
+export { fetchRateLimitDetails } from "./GitHubRateLimitApi.js";
+
+// Remotes
+export { listGitHubRemotes } from "./GitHubRemotes.js";
