+++ vitest.config.ts
@@ -21,6 +21,7 @@ export default defineConfig({
       "src/**/*.{test,spec}.{js,ts,jsx,tsx}",
       "shared/**/*.{test,spec}.{js,ts}",
       "scripts/**/*.{test,spec}.{js,ts,mjs}",
+      "plugins/**/*.{test,spec}.{js,ts}",
     ],
     exclude: [
       "node_modules",
